+++ tools/onnx-graphsurgeon/CHANGELOG.md
@@ -7,8 +7,7 @@ Dates are in YYYY-MM-DD format.
 - Adds `BaseImporter` and `OnnxImporter` classes.
 - Adds support for importing initializers in the `OnnxImporter`
 - Adds `VariableTensor` and `ConstantTensor`
-- Adds shape inference during ONNX Graph import.
-- Consolidates inputs/outputs of Nodes/Tensors. Now, inputs/outputs should only ever be added to `Node`s.
+- Consolidates inputs/outputs of Nodes/Tensors. Now, inputs/outputs should generally only be added to `Node`s.
 - Adds `OnnxExporter` to export `Graph` to `onnx.GraphProto`
 - Adds `OnnxExporter` and `OnnxImporter` to public imports
 - Adds `toposort` function to `Graph`, which will topologically sort it.